- libprojectM: Ogg player updates for example applications (smaller audio buffers, better PCM callback response) - libprojectM: Added libogc style u32 audio buffer support for PCM - libprojectM: More work on openGL <-> GX wrapper

+++ branches/newgui/libs/libprojectM/example/source/oggplayer.h
@@ -57,7 +57,7 @@ typedef union _dword {
 } dword;
 
 // Event callbacks
-typedef void (*ogg_cb_pcmdecode) (s16 *pcm, u32 samples);
+typedef void (*ogg_cb_pcmdecode) (u32 *pcm, u32 samples);
 
 // Ogg player routines
 bool oggPlay (const char *filepath, ogg_cb_pcmdecode onpcmdecode);
